# Optimise merge operation