It is possible to get ride of boundary check in the Merge operation


```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/hyperloglog
         │   old.txt   │               new.txt               │
         │   sec/op    │   sec/op     vs base                │
Merge-10   977.3n ± 1%   661.1n ± 1%  -32.35% (p=0.000 n=12)

         │  old.txt   │            new.txt             │
         │    B/op    │    B/op     vs base            │
Merge-10   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=12) ¹
¹ all samples are equal

         │  old.txt   │            new.txt             │
         │ allocs/op  │ allocs/op   vs base            │
Merge-10   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=12) ¹
```